Allow users to specify a screenshot directory
